All message validation result types moved to namespace `Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe.Validation`
