+++ src/Oriflame.RedisMessaging.ReliableDelivery/Oriflame.RedisMessaging.ReliableDelivery.csproj
@@ -13,6 +13,7 @@
 
   <PropertyGroup>
     <CodeAnalysisRuleSet>Ori.Common.Clustering.Events.Redis.ruleset</CodeAnalysisRuleSet>
+    <LangVersion>7.2</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
