--- src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/IMessageValidationResult.cs
@@ -1,4 +1,0 @@
-﻿namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe
-{
-    public interface IMessageValidationResult { }
-}
