+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/Validation/AlreadyProcessedValidationResult.cs
@@ -1,0 +1,11 @@
+﻿namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe.Validation
+{
+    public sealed class AlreadyProcessedValidationResult : IMessageValidationResult
+    {
+        public static readonly AlreadyProcessedValidationResult Instance  = new AlreadyProcessedValidationResult();
+
+        private AlreadyProcessedValidationResult()
+        {
+        }
+    }
+}
