+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/Validation/IMessageValidationResult.cs
@@ -1,0 +1,4 @@
+﻿namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe.Validation
+{
+    public interface IMessageValidationResult { }
+}
