+++ src/Oriflame.RedisMessaging.ReliableDelivery/Subscribe/Validation/SuccessValidationResult.cs
@@ -1,0 +1,11 @@
+﻿namespace Oriflame.RedisMessaging.ReliableDelivery.Subscribe.Validation
+{
+    public sealed class SuccessValidationResult : IMessageValidationResult
+    {
+        public static readonly SuccessValidationResult Instance  = new SuccessValidationResult();
+
+        private SuccessValidationResult()
+        {
+        }
+    }
+}
