+++ test/Oriflame.RedisMessaging.ReliableDelivery.Tests/Subscribe/Validation/MessageValidatorTest.cs
@@ -1,7 +1,8 @@
 ﻿using Oriflame.RedisMessaging.ReliableDelivery.Subscribe;
+using Oriflame.RedisMessaging.ReliableDelivery.Subscribe.Validation;
 using Xunit;
 
-namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe
+namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe.Validation
 {
     public class MessageValidatorTest
     {
@@ -13,9 +14,9 @@ namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe
 
             // act & assert
             var message1 = new Message(2, "test-message");
-            Assert.Same(MessageValidationResult.Success,  validator.Validate(message1));
+            Assert.IsType<SuccessValidationResult>(validator.Validate(message1));
             var message2 = new Message(3, "test-message");
-            Assert.Same(MessageValidationResult.Success, validator.Validate(message2));
+            Assert.IsType<SuccessValidationResult>(validator.Validate(message2));
         }
 
         [Fact]
@@ -26,7 +27,7 @@ namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe
 
             // act & assert
             var message1 = new Message(1, "test-message");
-            Assert.Same(MessageValidationResult.Success, validator.Validate(message1));
+            Assert.IsType<SuccessValidationResult>(validator.Validate(message1));
 
             var message2 = new Message(3, "test-message");
             var failureResult = validator.Validate(message2);
@@ -43,8 +44,8 @@ namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe
             var message1 = new Message(1, "test-message");
 
             // act & assert
-            Assert.Same(MessageValidationResult.Success, validator.Validate(message1));
-            Assert.Same(MessageValidationResult.MessageAgain, validator.Validate(message1));
+            Assert.IsType<SuccessValidationResult>(validator.Validate(message1));
+            Assert.IsType<AlreadyProcessedValidationResult>(validator.Validate(message1));
         }
 
         [Fact]
@@ -55,7 +56,7 @@ namespace Oriflame.RedisMessaging.ReliableDelivery.Tests.Subscribe
 
             // act & assert
             var message = new Message(2000, "test-message"); // some large message ID
-            Assert.Same(MessageValidationResult.Success, validator.Validate(message));
+            Assert.IsType<SuccessValidationResult>(validator.Validate(message));
         }
     }
 }
